+++ kernel/src/print.rs
@@ -13,15 +13,30 @@ macro_rules! println {
     ($($arg:tt)*) => ($crate::print!("{}\n", format_args!($($arg)*)));
 }
 
-// Stub implementations for other architectures
-#[cfg(not(target_arch = "x86_64"))]
+// AArch64 implementation using serial port
+#[cfg(target_arch = "aarch64")]
 #[macro_export]
 macro_rules! print {
-    ($($arg:tt)*) => {};
+    ($($arg:tt)*) => ($crate::serial::_serial_print(format_args!($($arg)*)));
 }
 
-#[cfg(not(target_arch = "x86_64"))]
+#[cfg(target_arch = "aarch64")]
 #[macro_export]
 macro_rules! println {
-    ($($arg:tt)*) => {};
+    () => ($crate::print!("\n"));
+    ($($arg:tt)*) => ($crate::print!("{}\n", format_args!($($arg)*)));
+}
+
+// RISC-V implementation using serial port
+#[cfg(target_arch = "riscv64")]
+#[macro_export]
+macro_rules! print {
+    ($($arg:tt)*) => ($crate::serial::_serial_print(format_args!($($arg)*)));
+}
+
+#[cfg(target_arch = "riscv64")]
+#[macro_export]
+macro_rules! println {
+    () => ($crate::print!("\n"));
+    ($($arg:tt)*) => ($crate::print!("{}\n", format_args!($($arg)*)));
 }